+++ R/batchtools_openlava.R
@@ -50,10 +50,11 @@ BatchtoolsOpenLavaFutureBackend <- function(...) {
 #'
 #' f <- future({
 #'   data.frame(
-#'     hostname = Sys.info()[["nodename"]],
-#'           os = Sys.info()[["sysname"]],
-#'        cores = unname(parallelly::availableCores()),
-#'      modules = Sys.getenv("LOADEDMODULES")
+#'      hostname = Sys.info()[["nodename"]],
+#'            os = Sys.info()[["sysname"]],
+#'     osVersion = utils::osVersion,
+#'         cores = unname(parallelly::availableCores()),
+#'       modules = Sys.getenv("LOADEDMODULES")
 #'   )
 #' })
 #' info <- value(f)