+++ R/batchtools_sge.R
@@ -65,10 +65,11 @@ print.BatchtoolsSGEFutureBackend <- function(x, ...) {
 #'
 #' f <- future({
 #'   data.frame(
-#'     hostname = Sys.info()[["nodename"]],
-#'           os = Sys.info()[["sysname"]],
-#'        cores = unname(parallelly::availableCores()),
-#'      modules = Sys.getenv("LOADEDMODULES")
+#'      hostname = Sys.info()[["nodename"]],
+#'            os = Sys.info()[["sysname"]],
+#'     osVersion = utils::osVersion,
+#'         cores = unname(parallelly::availableCores()),
+#'       modules = Sys.getenv("LOADEDMODULES")
 #'   )
 #' })
 #' info <- value(f)